requirements.txt
requests
tqdm
pandas
matplotlib
fpdf
python-dotenv